[PCI refactor] Move PCI related files to pci subdirectory
Issue #417
