+++ kernel/drivers/pci/mod.rs
@@ -1,0 +1,4 @@
+pub mod config;
+mod init;
+
+pub use drivers::pci::init::pci_init;
