+++ kernel/network/intel8254x.rs
@@ -8,7 +8,7 @@ use core::ptr;
 
 use common::{debug, memory};
 
-use drivers::pciconfig::PciConfig;
+use drivers::pci::config::PciConfig;
 
 use network::common::*;
 use network::scheme::*;
