+++ kernel/network/rtl8139.rs
@@ -9,7 +9,7 @@ use core::ptr;
 
 use common::{debug, memory};
 
-use drivers::pciconfig::PciConfig;
+use drivers::pci::config::PciConfig;
 use drivers::pio::*;
 use drivers::mmio::Mmio;
