+++ kernel/usb/xhci.rs
@@ -1,4 +1,4 @@
-use drivers::pciconfig::PciConfig;
+use drivers::pci::config::PciConfig;
 
 use core::mem::size_of;
